Extract `scarb-ui` crate
commit-id:d43533b4
