+++ Cargo.toml
@@ -8,6 +8,7 @@ members = [
     "utils/create-output-dir",
     "utils/scarb-build-metadata",
     "utils/scarb-test-support",
+    "utils/scarb-ui",
     "utils/test-for-each-example",
     "xtask",
 ]
