+++ scarb/src/lib.rs
@@ -6,6 +6,8 @@
 #![deny(rustdoc::private_intra_doc_links)]
 #![warn(rust_2018_idioms)]
 
+pub use subcommands::EXTERNAL_CMD_PREFIX;
+
 pub mod compiler;
 pub mod core;
 pub mod flock;
@@ -16,12 +18,9 @@ pub mod process;
 mod resolver;
 mod sources;
 mod subcommands;
-pub mod ui;
 pub mod version;
 
 pub const SCARB_ENV: &str = "SCARB";
 pub const MANIFEST_FILE_NAME: &str = "Scarb.toml";
 pub const DEFAULT_SOURCE_PATH: &str = "src/lib.cairo";
 pub const DEFAULT_TARGET_DIR_NAME: &str = "target";
-
-pub use subcommands::EXTERNAL_CMD_PREFIX;
