+++ utils/scarb-ui/src/lib.rs
@@ -1,8 +1,8 @@
-use anyhow::{bail, Result};
 use clap::ValueEnum;
-use std::env;
-use std::fmt::Display;
-use std::str::FromStr;
+pub use indicatif::{
+    BinaryBytes, DecimalBytes, FormattedDuration, HumanBytes, HumanCount, HumanDuration,
+    HumanFloatCount,
+};
 
 pub use machine::*;
 pub use message::*;
@@ -10,6 +10,7 @@ pub use spinner::*;
 pub use status::*;
 pub use typed::*;
 pub use value::*;
+pub use verbosity::*;
 pub use widget::*;
 
 mod machine;
@@ -18,50 +19,8 @@ mod spinner;
 mod status;
 mod typed;
 mod value;
+mod verbosity;
 mod widget;
-
-/// The requested verbosity of output.
-///
-/// # Ordering
-/// [`Verbosity::Quiet`] < [`Verbosity::Normal`] < [`Verbosity::Verbose`]
-#[derive(Copy, Clone, Debug, Default, Eq, PartialEq, Ord, PartialOrd)]
-pub enum Verbosity {
-    Quiet,
-    #[default]
-    Normal,
-    Verbose,
-}
-
-impl Display for Verbosity {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Quiet => write!(f, "quiet"),
-            Self::Normal => write!(f, "normal"),
-            Self::Verbose => write!(f, "verbose"),
-        }
-    }
-}
-
-impl FromStr for Verbosity {
-    type Err = anyhow::Error;
-
-    fn from_str(s: &str) -> Result<Self> {
-        match s {
-            "quiet" => Ok(Verbosity::Quiet),
-            "normal" => Ok(Verbosity::Normal),
-            "verbose" => Ok(Verbosity::Verbose),
-            "" => bail!("empty string cannot be used as verbosity level"),
-            _ => bail!("invalid verbosity level: {s}"),
-        }
-    }
-}
-
-impl Verbosity {
-    pub fn from_env_var(env_var_name: &str) -> Result<Self> {
-        let env_var = env::var(env_var_name)?;
-        Self::from_str(env_var.as_str())
-    }
-}
 
 /// The requested format of output (either textual or JSON).
 #[derive(Copy, Clone, Debug, Default, Eq, PartialEq, ValueEnum)]
@@ -139,39 +98,9 @@ impl Ui {
             OutputFormat::Json => message.print_json(),
         }
     }
-}
 
-#[cfg(test)]
-mod tests {
-    use super::Verbosity;
-    use std::env;
-
-    #[test]
-    fn verbosity_ord() {
-        use Verbosity::*;
-        assert!(Quiet < Normal);
-        assert!(Normal < Verbose);
-    }
-
-    #[test]
-    fn verbosity_from_str() {
-        use Verbosity::*;
-        assert_eq!(Quiet.to_string().parse::<Verbosity>().unwrap(), Quiet);
-        assert_eq!(Normal.to_string().parse::<Verbosity>().unwrap(), Normal);
-        assert_eq!(Verbose.to_string().parse::<Verbosity>().unwrap(), Verbose);
-    }
-
-    #[test]
-    fn verbosity_from_env_var() {
-        use Verbosity::*;
-        env::set_var("SOME_ENV_VAR", "quiet");
-        assert_eq!(Verbosity::from_env_var("SOME_ENV_VAR").unwrap(), Quiet);
-        env::set_var("SOME_ENV_VAR", "verbose");
-        assert_eq!(Verbosity::from_env_var("SOME_ENV_VAR").unwrap(), Verbose);
-
-        assert!(matches!(
-            Verbosity::from_env_var("SOME_ENV_VAR_THAT_DOESNT_EXIST"),
-            Err(_)
-        ));
+    /// Forces colorization on or off for stdout.
+    pub fn force_colors_enabled(&self, enable: bool) {
+        console::set_colors_enabled(enable);
     }
 }
